Rename the dnb_api module
This is to avoid confusion between the local api module and the dnb api module, which is now called dnb_direct_plus
